Fixed autocomplete form attribute
Prior to this commit, all input fields generated a autocomplete attribute with a value equal to the name of the field.

After this commit, a default autocomplete attribute is not generated unless the name of the field matches one of the valid autocomplete values.

Autocomplete can now be disabled on the entire form by calling `->autocomplete('off')`, the default being 'on'.

See: https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#autofill